Guard against RateLimiter failures causing notifications to fail
Partial resolution to issue #39.  We now guard against the possibility of the
rate limiter implementation throwing an exception, which could result in important
notifications not being sent to users. If some exception occurs during the course
of our checking the rate limiter status, we now report and log the error but
continue to send the notification in any event.

This may result in more notifications being sent than intended, but that failure
mode is preferable to notifications not being sent at all.